+++ src/components/SegmentedControl.tsx
@@ -150,7 +150,6 @@ const styles = StyleSheet.create({
     container: {
         flexDirection: 'row',
         alignItems: 'center',
-        flexWrap: 'wrap',
     },
     item: {
         flexShrink: 1,